requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
kafka_python